[tests] Small refactoring
